Be able to place ships on the board
That took a good while. In the end, I used callbacks to call the same
function for the entire array of ships. Not entirely happy with the
implementation. There must be a better way to implement changing ships.
On the plus side, my implementation supports any number of ships, which
will come in handy when I'll be implementing settings/options.