Improve icons and cursor types
In the ship placement modal, icons that cannot be clicked turn grey
and change cursor to default. Also, board squares that are not
clickable or are no longer clickable have their cursor set back to
default. Finally, I moved some functions from DOM.modals to DOM because
they are not necessary there and DOM.modals is already big.